backend/requirements.txt
# Django Core
Django>=4.2
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
django-filter>=23.5

# Database (optional - for PostgreSQL)
# psycopg2-binary>=2.9.9  # Uncomment if using PostgreSQL

# PDF Processing
PyPDF2>=3.0.1
pypdfium2>=5.0.0  # Primary PDF renderer (works on Windows without dependencies)
pdf2image>=1.16.3  # Alternative (requires poppler)
Pillow>=10.1.0

# File Storage (AWS S3 optional)
boto3>=1.34.0
django-storages>=1.14.2

# Task Queue (for background processing)
celery>=5.3.4
redis>=5.0.1

# API Documentation
drf-yasg>=1.21.7

# Environment
python-dotenv>=1.0.0

# Production Server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Testing
pytest>=7.4.3
pytest-django>=4.7.0

# Development
django-debug-toolbar>=4.2.0
